Move AAX UTF table reading to its own file for clarity
